+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - **Badge** add prop `type`
+- **Spinner** add prop `color`
+- **Spinner** add default color of `.c-action-primary`
 
 ### Changed
 
@@ -16,7 +18,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **[BREAKING]** **Alert** Remove value `info` from `type` prop
 - **[BREAKING]** Icons default color is now `currentColor`, which is the current text color
 - **[BREAKING]** **Spinner** remove prop `secondary`
-- **[BREAKING]** **Spinner** add prop `color`
 
 ### Fixed
 - Spinner proptypes